Fixed package structure in examples
