--- example-graphql-subscription/src/main/java/com/embedler/moon/graphql/boot/SubscriptionSampleApplication.java
@@ -1,13 +1,0 @@
-package com.embedler.moon.graphql.boot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SubscriptionSampleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SubscriptionSampleApplication.class, args);
-    }
-
-}
