--- example-graphql-subscription/src/main/java/com/embedler/moon/graphql/boot/resolvers/Subscription.java
@@ -1,23 +1,0 @@
-package com.embedler.moon.graphql.boot.resolvers;
-
-import com.coxautodev.graphql.tools.GraphQLSubscriptionResolver;
-import com.embedler.moon.graphql.boot.publishers.StockTickerPublisher;
-import org.reactivestreams.Publisher;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-class Subscription implements GraphQLSubscriptionResolver {
-
-    private StockTickerPublisher stockTickerPublisher;
-
-    Subscription(StockTickerPublisher stockTickerPublisher) {
-        this.stockTickerPublisher = stockTickerPublisher;
-    }
-
-    Publisher<StockPriceUpdate> stockQuotes(List<String> stockCodes) {
-        return stockTickerPublisher.getPublisher();
-    }
-
-}
