+++ example-graphql-subscription/src/main/java/com/oembedler/moon/graphql/boot/SubscriptionSampleApplication.java
@@ -1,0 +1,13 @@
+package com.oembedler.moon.graphql.boot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SubscriptionSampleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SubscriptionSampleApplication.class, args);
+    }
+
+}
