--- example-graphql-tools/src/main/java/com/embedler/moon/graphql/boot/GraphQLToolsSampleApplication.java
@@ -1,13 +1,0 @@
-package com.embedler.moon.graphql.boot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GraphQLToolsSampleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GraphQLToolsSampleApplication.class, args);
-    }
-
-}
