--- example-graphql-tools/src/main/java/com/embedler/moon/graphql/boot/resolvers/Comment.java
@@ -1,8 +1,0 @@
-package com.embedler.moon.graphql.boot.resolvers;
-
-class Comment {
-
-    private Long id;
-    private String description;
-
-}
