--- example-graphql-tools/src/main/java/com/embedler/moon/graphql/boot/resolvers/Post.java
@@ -1,11 +1,0 @@
-package com.embedler.moon.graphql.boot.resolvers;
-
-class Post {
-
-    private Long id;
-
-    Post(Long id) {
-        this.id = id;
-    }
-
-}
