--- example-graphql-tools/src/main/java/com/embedler/moon/graphql/boot/resolvers/PostResolver.java
@@ -1,16 +1,0 @@
-package com.embedler.moon.graphql.boot.resolvers;
-
-import com.coxautodev.graphql.tools.GraphQLResolver;
-import org.springframework.stereotype.Component;
-
-import java.util.Collections;
-import java.util.List;
-
-@Component
-class PostResolver implements GraphQLResolver<Post> {
-
-    public List<Comment> getComments(Post post) {
-        return Collections.emptyList();
-    }
-
-}
