--- example-graphql-tools/src/main/java/com/embedler/moon/graphql/boot/resolvers/Query.java
@@ -1,13 +1,0 @@
-package com.embedler.moon.graphql.boot.resolvers;
-
-import com.coxautodev.graphql.tools.GraphQLQueryResolver;
-import org.springframework.stereotype.Component;
-
-@Component
-class Query implements GraphQLQueryResolver {
-
-    Post getPost(Long id) {
-        return new Post(id);
-    }
-
-}
