+++ example-graphql-tools/src/main/java/com/oembedler/moon/graphql/boot/GraphQLToolsSampleApplication.java
@@ -1,0 +1,13 @@
+package com.oembedler.moon.graphql.boot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GraphQLToolsSampleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GraphQLToolsSampleApplication.class, args);
+    }
+
+}
