+++ example-graphql-tools/src/main/java/com/oembedler/moon/graphql/boot/resolvers/Comment.java
@@ -1,0 +1,8 @@
+package com.oembedler.moon.graphql.boot.resolvers;
+
+class Comment {
+
+    private Long id;
+    private String description;
+
+}
