+++ example-graphql-tools/src/main/java/com/oembedler/moon/graphql/boot/resolvers/Post.java
@@ -1,0 +1,11 @@
+package com.oembedler.moon.graphql.boot.resolvers;
+
+class Post {
+
+    private Long id;
+
+    Post(Long id) {
+        this.id = id;
+    }
+
+}
